apps/embedding/pyproject.toml
[project]
name = "embedding"
version = "0.1.0"
description = "Reads a JSONL with chunks from S3 and generate the embedding for each chunk in a vector database (Qdrant)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.54",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "utils",
]

[tool.uv.sources]
utils = { path = "../utils", editable = true }
